pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.264"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
